rename: 아이콘 폴더 위치 변경
