+++ src/assets/icons/index.ts
@@ -1,0 +1,33 @@
+import PencilIcon from './pencil.svg?react'
+import KakaoIcon from './kakao.svg?react'
+import ArrowLeftIcon from './arrow-left.svg?react'
+import ArrowUpIcon from './arrow-up.svg?react'
+import HamburgerIcon from './hamburger.svg?react'
+import SettingIcon from './setting.svg?react'
+import ShareIcon from './share.svg?react'
+import QuestionIcon from './question.svg?react'
+import CheckIcon from './check.svg?react'
+import MessageIcon from './message.svg?react'
+import ThumbUpIcon from './thumb-up.svg?react'
+import SearchIcon from './search.svg?react'
+import CancelIcon from './cancel.svg?react'
+import ArrowSmallUpIcon from './arrow-small-up.svg?react'
+import LogoIcon from './logo.svg?react'
+
+export {
+  PencilIcon,
+  KakaoIcon,
+  ArrowLeftIcon,
+  ArrowUpIcon,
+  HamburgerIcon,
+  SettingIcon,
+  ShareIcon,
+  QuestionIcon,
+  CheckIcon,
+  MessageIcon,
+  ThumbUpIcon,
+  SearchIcon,
+  CancelIcon,
+  ArrowSmallUpIcon,
+  LogoIcon,
+}
