--- src/assets/index.ts
@@ -1,33 +1,0 @@
-import PencilIcon from '@/assets/pencil.svg?react'
-import KakaoIcon from '@/assets/kakao.svg?react'
-import ArrowLeftIcon from '@/assets/arrow-left.svg?react'
-import ArrowUpIcon from '@/assets/arrow-up.svg?react'
-import HamburgerIcon from '@/assets/hamburger.svg?react'
-import SettingIcon from '@/assets/setting.svg?react'
-import ShareIcon from '@/assets/share.svg?react'
-import QuestionIcon from '@/assets/question.svg?react'
-import CheckIcon from '@/assets/check.svg?react'
-import MessageIcon from '@/assets/message.svg?react'
-import ThumbUpIcon from '@/assets/thumb-up.svg?react'
-import SearchIcon from '@/assets/search.svg?react'
-import CancelIcon from '@/assets/cancel.svg?react'
-import ArrowSmallUpIcon from '@/assets/arrow-small-up.svg?react'
-import LogoIcon from '@/assets/logo.svg?react'
-
-export {
-  PencilIcon,
-  KakaoIcon,
-  ArrowLeftIcon,
-  ArrowUpIcon,
-  HamburgerIcon,
-  SettingIcon,
-  ShareIcon,
-  QuestionIcon,
-  CheckIcon,
-  MessageIcon,
-  ThumbUpIcon,
-  SearchIcon,
-  CancelIcon,
-  ArrowSmallUpIcon,
-  LogoIcon,
-}
