+++ src/components/Dropdown.tsx
@@ -1,4 +1,4 @@
-import { ArrowSmallUpIcon } from '@/assets'
+import { ArrowSmallUpIcon } from '@/assets/icons'
 import { useEffect, useRef, useState } from 'react'
 import { twMerge } from 'tailwind-merge'
 
@@ -42,7 +42,7 @@ export default function Dropdown({ options, selected, className, handleSelected 
         <ArrowSmallUpIcon className="h-6 w-6" />
       </button>
       {isOpen && (
-        <ul className="label-medium shadow-dropdown absolute bottom-0 top-full mt-2 h-fit w-full rounded-lg bg-neutral-80 py-2 text-white">
+        <ul className="label-medium absolute bottom-0 top-full mt-2 h-fit w-full rounded-lg bg-neutral-80 py-2 text-white shadow-dropdown">
           {options.map((option) => (
             <li key={option} className="hover:bg-neutral-70">
               <button
